# docs/swagger.yaml
basePath: /api/v1
definitions:
  image.ImageURLResponse:
    properties:
      url:
        type: string
    type: object
  image.UploadImageResponse:
    properties:
      image_id:
        type: string
      url:
        type: string
    type: object
...
  models.ErrorCode:
    enum:
    - INVALID_REQUEST
    - MISSING_FIELDS
    - INVALID_EMAIL
    - INVALID_PASSWORD
    - INVALID_LOGIN
    - USER_EXISTS
    - EMAIL_EXISTS
    - LOGIN_EXISTS
    - WEAK_PASSWORD
    - USER_NOT_FOUND
    - INVALID_CREDENTIALS
    - ACCOUNT_LOCKED
    - TOKEN_EXPIRED
    - TOKEN_INVALID
    - TOKEN_MISSING
    - UNAUTHORIZED
    - FORBIDDEN
    - ACCESS_DENIED
    - RESOURCE_NOT_FOUND
    - RESOURCE_EXISTS
    - RESOURCE_CONFLICT
    - BUDGET_NOT_FOUND
    - ACCOUNT_NOT_FOUND
    - OPERATION_NOT_FOUND
    - INVALID_AMOUNT
    - INVALID_CURRENCY
    - INVALID_DATE
    - INVALID_PERIOD
    - INTERNAL_ERROR
    - SERVICE_UNAVAILABLE
    - DATABASE_ERROR
    type: string
    x-enum-varnames:
    - ErrCodeInvalidRequest
    - ErrCodeMissingFields
    - ErrCodeInvalidEmail
    - ErrCodeInvalidPassword
    - ErrCodeInvalidLogin
    - ErrCodeUserExists
    - ErrCodeEmailExists
    - ErrCodeLoginExists
    - ErrCodeWeakPassword
    - ErrCodeUserNotFound
    - ErrCodeInvalidCredentials
    - ErrCodeAccountLocked
    - ErrCodeTokenExpired
    - ErrCodeTokenInvalid
    - ErrCodeTokenMissing
    - ErrCodeUnauthorized
    - ErrCodeForbidden
    - ErrCodeAccessDenied
    - ErrCodeResourceNotFound
    - ErrCodeResourceExists
    - ErrCodeResourceConflict
    - ErrCodeBudgetNotFound
    - ErrCodeAccountNotFound
    - ErrCodeTransactionNotFound
    - ErrCodeInvalidAmount
    - ErrCodeInvalidCurrency
    - ErrCodeInvalidDate
    - ErrCodeInvalidPeriod
    - ErrCodeInternalError
    - ErrCodeServiceUnavailable
    - ErrCodeDatabaseError
  models.ErrorResponse:
    properties:
      code:
        $ref: '#/definitions/models.ErrorCode'
      details:
        type: string
      error:
        type: string
      field:
        type: string
      timestamp:
        type: string
    type: object
...
  models.ProfileResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      login:
        type: string
      logo_hashed_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для управления личными финансами
  title: VKarmane API
  version: "1.0"
paths:
...
  /images:
    delete:
      description: Удаляет изображение из хранилища MinIO по image_id
      parameters:
      - description: Идентификатор изображения (SHA256 hash)
        in: query
        name: image_id
        required: true
        type: string
      responses:
        "204":
          description: Изображение успешно удалено
        "400":
          description: Некорректные данные (INVALID_REQUEST, MISSING_IMAGE_ID)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Требуется аутентификация (UNAUTHORIZED, TOKEN_MISSING, TOKEN_INVALID,
            TOKEN_EXPIRED)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Изображение не найдено (RESOURCE_NOT_FOUND)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера (INTERNAL_ERROR)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удаление изображения
      tags:
      - images
    get:
      description: Выполняет редирект на presigned URL изображения по image_id. Не
        требует аутентификации
      parameters:
      - description: Идентификатор изображения (SHA256 hash)
        in: query
        name: image_id
        required: true
        type: string
      responses:
        "302":
          description: Редирект на изображение
        "400":
          description: Некорректные данные (INVALID_REQUEST, MISSING_IMAGE_ID)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Изображение не найдено (RESOURCE_NOT_FOUND)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера (INTERNAL_ERROR)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получение изображения (редирект)
      tags:
      - images
  /images/upload:
    post:
      consumes:
      - multipart/form-data
      description: Загружает изображение в хранилище MinIO и возвращает image_id и
        URL
      parameters:
      - description: Изображение (jpg, jpeg, png, gif, webp)
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Изображение успешно загружено
          schema:
            $ref: '#/definitions/image.UploadImageResponse'
        "400":
          description: Некорректные данные (INVALID_REQUEST, INVALID_IMAGE_FORMAT)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Требуется аутентификация (UNAUTHORIZED, TOKEN_MISSING, TOKEN_INVALID,
            TOKEN_EXPIRED)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера (INTERNAL_ERROR)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Загрузка изображения
      tags:
      - images
  /images/url:
    get:
      description: Возвращает presigned URL для доступа к изображению по image_id
      parameters:
      - description: Идентификатор изображения (SHA256 hash)
        in: query
        name: image_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: URL изображения
          schema:
            $ref: '#/definitions/image.ImageURLResponse'
        "400":
          description: Некорректные данные (INVALID_REQUEST, MISSING_IMAGE_ID)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Требуется аутентификация (UNAUTHORIZED, TOKEN_MISSING, TOKEN_INVALID,
            TOKEN_EXPIRED)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Изображение не найдено (RESOURCE_NOT_FOUND)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера (INTERNAL_ERROR)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение URL изображения
      tags:
      - images
...
  /profile:
    get:
      description: Возвращает информацию о профиле текущего пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Профиль пользователя
          schema:
            $ref: '#/definitions/models.ProfileResponse'
        "401":
          description: Требуется аутентификация (UNAUTHORIZED, TOKEN_MISSING, TOKEN_INVALID,
            TOKEN_EXPIRED)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера (INTERNAL_ERROR, DATABASE_ERROR)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получение профиля пользователя
      tags:
      - profile
  /profile/edit:
    put:
      consumes:
      - multipart/form-data
      description: Обновляет информацию о профиле текущего пользователя. Поддерживает
        multipart/form-data с опциональным полем avatar для загрузки аватарки
      parameters:
      - description: Имя
        in: formData
        name: first_name
        required: true
        type: string
      - description: Фамилия
        in: formData
        name: last_name
        required: true
        type: string
      - description: Email
        in: formData
        name: email
        required: true
        type: string
      - description: Аватарка (опционально)
        in: formData
        name: avatar
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Обновленный профиль пользователя
          schema:
            $ref: '#/definitions/models.ProfileResponse'
        "400":
          description: Некорректные данные (INVALID_REQUEST, MISSING_FIELDS)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Требуется аутентификация (UNAUTHORIZED, TOKEN_MISSING, TOKEN_INVALID,
            TOKEN_EXPIRED)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Email уже используется (RESOURCE_EXISTS)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера (INTERNAL_ERROR, DATABASE_ERROR)
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновление профиля пользователя
      tags:
      - profile
securityDefinitions:
  ApiKeyAuth:
    description: JWT токен в HTTP-only cookie
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"